Add FCM module missing unit tests (#SVC-3995)
