+++ .rbenv-version
@@ -1,0 +1,2 @@
+2.2.6
+
