+++ .rvmrc
@@ -1,0 +1,3 @@
+rvm_trust_rvmrcs_flag=1
+rvm use 2.2.6@applicationservice --create
+
